♻️ refactor: 폴더 위치 providers로 변경
